+++ src/main/java/com/batch/GUI/InitialWindow/SubWindows/HelpWindow.java
@@ -1,5 +1,3 @@
-
-
 package com.batch.GUI.InitialWindow.SubWindows;
 
 import javafx.geometry.Insets;
@@ -20,15 +18,15 @@ public class HelpWindow extends Stage {
         Label label = new Label("Mixing platform V2.1\nSystem by CIRCLE\n");
         label.setTextAlignment(TextAlignment.CENTER);
         label.setAlignment(Pos.CENTER);
-        label.setFont(Font.font("Arial", FontWeight.BOLD, FontPosture.REGULAR, (double)15.0F));
-        label.setPadding(new Insets((double)10.0F));
+        label.setFont(Font.font("Arial", FontWeight.BOLD, FontPosture.REGULAR, 15.0F));
+        label.setPadding(new Insets(10.0F));
         label.prefWidthProperty().bind(this.widthProperty());
         vBox.getChildren().add(label);
-        vBox.setPadding(new Insets((double)20.0F));
+        vBox.setPadding(new Insets(20.0F));
         Scene scene = new Scene(vBox);
         this.setScene(scene);
         this.setTitle("About system");
-        this.setWidth((double)500.0F);
+        this.setWidth(500.0F);
         this.setResizable(false);
         this.setAlwaysOnTop(true);
         this.initStyle(StageStyle.UTILITY);
